Remove explicit dependency on Babel
Bit-rot between my last activity in this project (3 yrs, 5 months ago)
and installing the deps yesterday led to a totally broken Babel install
which blows up on invoking Gulp. The project is so simple it doesn't
need Babel, so remove it; in a subsequent commit I'll manually rewrite
it in ES5.